% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPal.R
\name{color_pal}
\alias{color_pal}
\title{Title}
\usage{
color_pal(palette, direction = 1)
}
\arguments{
\item{direction}{}
}
\value{

}
\description{
Title
}
\examples{
names(pal_list)
scales::show_col(color_pal("Reds")(5))
}
