% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPal.R
\name{scale_colors}
\alias{scale_colors}
\title{Title}
\usage{
scale_colors(palette, direction = 1, aesthetics = "colour", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
(d <- ggplot(dsamp, aes(carat, price)) +
    geom_point(aes(colour = clarity)))
d + scale_colour_brewer()
d + scale_colors("Viridis")
d + scale_colors("Viridis", aesthetics = "fill")
d + scale_colors("Reds")
d + scale_colors("Earth")
d + scale_colors("Portland")
}
